refactor(session): house the workspace provider vocabulary with the session vocabulary
SUPERSET_WORKSPACE_PROVIDER_ID, WorkspaceProviderId, and
isWorkspaceProviderId were defined in the superset package's vocabulary
subpath even though they describe the app-wide set of workspace-capable
providers and half of that set — CONDUCTOR_LOCAL_WORKSPACE_PROVIDER_ID
and every ProviderId — already lived in @sidecar/session. They now stand
beside PROVIDER_ID in packages/session/src/providers.ts, unchanged in
value, type, and guard; every importer reaches them through
@sidecar/session, the superset vocabulary subpath is gone, the settings
package drops its superset dependency, and the superset workspace test
stops importing the id from the desktop app's contracts.
