+++ apps/desktop/src/renderer/superset-sign-in-slot.tsx
@@ -1,4 +1,4 @@
-import { SUPERSET_WORKSPACE_PROVIDER_ID } from "@sidecar/superset/vocabulary";
+import { SUPERSET_WORKSPACE_PROVIDER_ID } from "@sidecar/session";
 import { useEffect, useRef, useState } from "react";
 import type { SupersetSignInSnapshot } from "#shared/contracts";
 import { CREDENTIAL_SOURCE, SUPERSET_SIGN_IN_STAGE } from "#shared/contracts";
