+++ apps/desktop/src/shared/contracts.ts
@@ -32,6 +32,7 @@ import type {
   SessionFilter,
   SessionIdentity,
   WorkspaceAgentSelection,
+  WorkspaceProviderId,
 } from "@sidecar/session";
 import type {
   AppSettingField,
@@ -42,7 +43,6 @@ import type {
   VoiceSource,
 } from "@sidecar/settings";
 import type { SupersetSignInSnapshot } from "@sidecar/superset/sign-in-stage";
-import type { WorkspaceProviderId } from "@sidecar/superset/vocabulary";
 import type {
   PanelFormFactor,
   Rectangle,
@@ -59,7 +59,13 @@ export {
   type AccountSnapshot,
 } from "@sidecar/account/snapshot";
 export type { AccountCalendar, ObservedAccountCalendars } from "@sidecar/calendar/observation";
-export { CLI_CONNECTION, type CliConnection } from "@sidecar/session";
+export {
+  CLI_CONNECTION,
+  type CliConnection,
+  isWorkspaceProviderId,
+  SUPERSET_WORKSPACE_PROVIDER_ID,
+  type WorkspaceProviderId,
+} from "@sidecar/session";
 export type {
   AppSettingField,
   AppSettingValue,
@@ -80,11 +86,6 @@ export {
   type SupersetOrganizationChoice,
   type SupersetSignInSnapshot,
 } from "@sidecar/superset/sign-in-stage";
-export {
-  isWorkspaceProviderId,
-  SUPERSET_WORKSPACE_PROVIDER_ID,
-  type WorkspaceProviderId,
-} from "@sidecar/superset/vocabulary";
 export type { WindowMode } from "@sidecar/surface";
 
 export function isAccountProvider(value: UnparsedWireValue): value is AccountProvider {
