+++ packages/session/src/index.ts
@@ -8,6 +8,7 @@ export {
   type HostedAgentId,
   isHostedAgentId,
   isProviderId,
+  isWorkspaceProviderId,
   maximumObservedWorkspaceProjects,
   maximumWorkspaceNameLength,
   normalizeObservedWorkspaceProjects,
@@ -28,11 +29,13 @@ export {
   type ProviderWorkspaceResult,
   type SessionProviderAdapter,
   SessionProviderAdapterBase,
+  SUPERSET_WORKSPACE_PROVIDER_ID,
   staleWorkspaceProjectDefaults,
   WORKSPACE_TASK_SUPPORT,
   type WorkspaceAgentModels,
   type WorkspaceAgentSelection,
   type WorkspaceProject,
+  type WorkspaceProviderId,
   type WorkspaceTaskSupport,
   workspaceNameText,
   workspaceProjectSelectionId,
