+++ packages/superset/src/workspaces.ts
@@ -16,10 +16,10 @@ import {
   SESSION_APPLICATION_SCOPE,
   SESSION_LOCATION,
   SESSION_STATUS,
+  SUPERSET_WORKSPACE_PROVIDER_ID,
 } from "@sidecar/session";
 import { type WireRecord, wireRecord } from "@sidecar/wire";
 import { SUPERSET_CONTROL_ID } from "./cli.js";
-import { SUPERSET_WORKSPACE_PROVIDER_ID } from "./vocabulary.js";
 
 const SUPERSET_AGENT_PROVIDER = {
   claude: PROVIDER_ID.CLAUDE_CODE,
